[android_webview] Update fixed layout size on page scale changes.
We should update the fixedLayoutSize even if the physical size of
the AwContents doesn't change. This is because the fixedLayoutSize
is directly derived from physical size and the pageScaleFactor, so
if either of these change the fixedLayoutSize should be updated.
Currently the AwLayoutSizer code assumes that every requestLayout
results in an onSizeChanged, but that's not always the case.

BUG=internal b/10917278
TEST=AndroidWebViewTest
R=joth@chromium.org

Review URL: https://codereview.chromium.org/26705002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@227932 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b8586e9375911cefa955293b8e01ef27221e41ce

+++ android_webview/javatests/src/org/chromium/android_webview/test/AndroidViewIntegrationTest.java
@@ -62,7 +62,7 @@ public class AndroidViewIntegrationTest extends AwTestBase {
         }
 
         @Override
-        public void onPageScaleChanged(double pageScaleFactor) {
+        public void onPageScaleChanged(float pageScaleFactor) {
             super.onPageScaleChanged(pageScaleFactor);
             mOnPageScaleChangedHelper.notifyCalled();
         }
